deployment/terraform/cloud_run_function_chat/requirements.txt
functions-framework>=3.0.0
botbuilder-core>=4.14.0
botbuilder-schema>=4.14.0
fastapi
uvicorn
requests==2.32.4
botbuilder-integration-aiohttp>=4.14.5
google-adk==1.5.0
yfinance==0.2.56
psutil==5.9.5
litellm==1.66.3
google-generativeai==0.8.5
python-dotenv==1.1.0
pdfminer
pdfminer.six
immutabledict==4.2.1
sqlglot==26.10.1
db-dtypes==1.4.2
regex==2024.11.6
tabulate==0.9.0
google-cloud-aiplatform[adk,agent-engines]>=1.94.0
absl-py==2.2.2
pydantic==2.11.3
toolbox-core
matplotlib
IPython
load_dotenv
pandas==2.2.3
deprecated==1.2.18
cloudpickle